Replacing ClientHandlerServer with ClientHandler
